+++ src/main/kotlin/it/czerwinski/intellij/wavefront/editor/gl/shaders/TexturedWireframeShader.kt
@@ -1,0 +1,46 @@
+/*
+ * Copyright 2020-2021 Slawomir Czerwinski
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package it.czerwinski.intellij.wavefront.editor.gl.shaders
+
+import graphics.glimpse.shaders.annotations.Attribute
+import graphics.glimpse.shaders.annotations.AttributeRole
+import graphics.glimpse.shaders.annotations.ShaderParams
+import graphics.glimpse.shaders.annotations.Uniform
+import graphics.glimpse.textures.Texture
+import graphics.glimpse.types.Mat4
+import graphics.glimpse.types.Vec4
+
+@ShaderParams(
+    attributes = [
+        Attribute(name = "aPos", role = AttributeRole.POSITIONS, vectorSize = 3),
+        Attribute(name = "aTexCoord", role = AttributeRole.TEX_COORDS, vectorSize = 2),
+    ]
+)
+data class TexturedWireframeShader(
+
+    @Uniform(name = "uMVPMat")
+    val mvpMatrix: Mat4,
+
+    @Uniform(name = "uPointSize")
+    val pointSize: Float = 1f,
+
+    @Uniform(name = "uColor")
+    val color: Vec4,
+
+    @Uniform(name = "uTexture")
+    val texture: Texture,
+)