+++ src/main/kotlin/it/czerwinski/intellij/wavefront/lang/Icons.kt
@@ -48,6 +48,7 @@ val OBJ_LINE_ICON: Icon = getIcon("/icons/objLineIcon.svg")
 val OBJ_POINT_ICON: Icon = getIcon("/icons/objPointIcon.svg")
 
 val OBJ_FACE_VERTEX_ICON: Icon = getIcon("/icons/objFaceVertexIcon.svg")
+val OBJ_LINE_VERTEX_ICON: Icon = OBJ_FACE_VERTEX_ICON
 
 val OBJ_SMOOTHING_ON_ICON: Icon = getIcon("/icons/objSmoothingOnIcon.svg")
 val OBJ_SMOOTHING_OFF_ICON: Icon = getIcon("/icons/objSmoothingOffIcon.svg")