+++ src/main/resources/shaders/textured_wireframe_fragment_shader.glsl
@@ -1,0 +1,12 @@
+#version 100
+
+precision mediump float;
+
+uniform vec4 uColor;
+uniform sampler2D uTexture;
+
+varying vec2 vTexCoord;
+
+void main() {
+    gl_FragColor = vec4(texture2D(uTexture, vTexCoord).rgb * uColor.rgb, uColor.a);
+}