+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ## [Unreleased]
 ### Added
+- Support for textured lines
 - Support for displacement textures in 3D preview
 - Cropping textures toggle action in 3D preview
 - Setting for cropping textures